Separando Historico em componente
