+++ src/components/ItemHistoricoDePedido/ItemHistoricoDePedido.scss
@@ -1,0 +1,168 @@
+.pedido-container{
+    widows: 100%;
+    max-width: 1180px;
+    padding: 0 30px;
+    margin: 32px auto;
+    
+}
+
+.pedido-container .content {
+    width: 100%;
+    padding: 30px;
+    background: #f0f0f5;
+    box-shadow: 0 0 100px rgba(0, 0 ,0 , 0.1);
+    border-radius: 8px;
+
+}
+
+
+.pedido-container h1 {
+    margin-top: 20px;
+    text-align: center;
+    margin-bottom: 24px;
+    color: #6C6C6C;
+}
+.pedido-container ul {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    grid-gap: 24px;
+    list-style: none;
+}
+.pedido-container ul li {
+    background: #FFF;
+    padding: 24px;
+    border-radius: 8px;
+    position: relative;
+}
+
+.pedido-container ul li strong {
+    display: block;
+    margin-bottom: 16px;
+    color: #41414d
+}
+
+.pedido-container ul li p + strong {
+    margin-top: 32px;
+}
+
+.pedido-container ul li p {
+    color : #737380;
+    line-height: 21px;
+    font-size: 16px;
+}
+
+
+
+
+
+
+
+
+
+
+
+/* Progress Tracker v2 */
+
+ol.progress-track[data-steps="2"] li {
+  width: 49%;
+}
+ol.progress-track[data-steps="3"] li {
+  width: 33%;
+}
+ol.progress-track[data-steps="4"] li {
+  width: 24%;
+}
+ol.progress-track[data-steps="5"] li {
+  width: 19%;
+}
+ol.progress-track[data-steps="6"] li {
+  width: 16%;
+}
+ol.progress-track[data-steps="7"] li {
+  width: 14%;
+}
+ol.progress-track[data-steps="8"] li {
+  width: 12%;
+}
+ol.progress-track[data-steps="9"] li {
+  width: 11%;
+}
+.progress-track {
+  width: 100%;
+  list-style: none;
+  list-style-image: none;
+  margin: 20px 0 20px 0;
+  padding: 0;
+  position: relative;
+  z-index: 0;
+}
+.progress-track li {
+  float: left;
+  text-align: center;
+  position: relative;
+}
+.progress-track .name {
+  display: block;
+  vertical-align: bottom;
+  text-align: center;
+  margin-bottom: 1em;
+  color: black;
+  opacity: 0.3;
+}
+.progress-track .step {
+  color: black;
+  border: 3px solid silver;
+  background-color: silver;
+  border-radius: 50%;
+  line-height: 2.1;
+  width: 2.2em;
+  height: 2.2em;
+  display: inline-block;
+  z-index: 0;
+}
+.progress-track .step span {
+  opacity: 0.3;
+}
+.progress-track .active .name,
+.progress-track .active .step span {
+  opacity: 1;
+}
+.progress-track .step:before {
+  content: "";
+  display: block;
+  background-color: silver;
+  height: 0.4em;
+  width: 50%;
+  position: absolute;
+  bottom: 0.6em;
+  left: 0;
+  z-index: -1;
+}
+.progress-track .step:after {
+  content: "";
+  display: block;
+  background-color: silver;
+  height: 0.4em;
+  width: 50%;
+  position: absolute;
+  bottom: 0.6em;
+  right: 0;
+  z-index: -1;
+}
+.progress-track li:first-of-type .step:before {
+  display: none;
+}
+.progress-track li:last-of-type .step:after {
+  display: none;
+}
+.progress-track .done .step,
+.progress-track .done .step:before,
+.progress-track .done .step:after,
+.progress-track .active .step,
+.progress-track .active .step:before {
+  background-color: yellowgreen;
+}
+.progress-track .done .step,
+.progress-track .active .step {
+  border: 3px solid yellowgreen;
+}
